requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
emoji==2.12.1
executing==2.0.1
fonttools==4.55.2
frozenlist==1.5.0
greenlet==3.1.1
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.3.11
langchain-community==0.3.11
langchain-core==0.3.24
langchain-groq==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.147
marshmallow==3.23.1
match==0.3.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
MouseInfo==0.1.3
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openai==0.28.0
orjson==3.10.12
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.2.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.2.0
pypdf==5.1.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.2
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==70.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
tenacity==8.5.0
tk==0.1.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
validators==0.34.0
wcwidth==0.2.13
wheel==0.43.0
yarl==1.18.3
youtube-transcript-api==0.6.3
